Finished Dialogue
The Guide/Dialogue is complete

You can define a scene in the menu
A scene has a location, text, and a bool for to show text or not.

When the player gets close to the guide it activates the scene and starts displaying the text

The player has to left click to move to the next bit of text. They are frozen through out the scene.

At the end of a scene, if there is another one, the guide will move to the next location and wait for the player.